requirements.txt
matplotlib~=3.5.1
tensorflow-gpu~=2.7.0
numpy~=1.24.4
graphviz~=2.38
pandas~=1.5.0
statsmodels~=0.14.1